requirements.txt
Flask==3.0.0
langchain==0.1.20
langchain-anthropic==0.1.6
anthropic==0.23.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-json-logger==2.0.7
docker==7.0.0
httpx==0.27.2
